+++ Physics/Serialization/QuantityInfo.cs
@@ -4,5 +4,5 @@ public class QuantityInfo
 {
     public double Amount { get; set; }
 
-    public Dictionary<string, int> Unit { get; set; }
+    public Dictionary<string, float> Unit { get; set; }
 }
